+++ src/main/java/absolutelyaya/ultracraft/block/mapping/ExplosionListenerBlockEntity.java
@@ -4,10 +4,12 @@ import absolutelyaya.ultracraft.ExplosionHandler;
 import absolutelyaya.ultracraft.damage.DamageSources;
 import absolutelyaya.ultracraft.registry.BlockEntityRegistry;
 import net.minecraft.block.BlockState;
+import net.minecraft.entity.Entity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
 import net.minecraft.text.Text;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Box;
 import org.joml.Vector4f;
 
 import java.util.ArrayList;
@@ -17,6 +19,7 @@ public class ExplosionListenerBlockEntity extends AbstractListenerBlockEntity
 {
 	static List<String> attributes = new ArrayList<>();
 	float explosionDamage = 0.01f, explosionRadius = 3f;
+	boolean justKnockback;
 	
 	public ExplosionListenerBlockEntity(BlockPos pos, BlockState state)
 	{
@@ -52,8 +55,18 @@ public class ExplosionListenerBlockEntity extends AbstractListenerBlockEntity
 	protected void onStateChanged(boolean newState)
 	{
 		if(newState && !world.isClient)
-			ExplosionHandler.explosion(null, world, getPos().toCenterPos(), DamageSources.get(world, DamageSources.EXPLOSION),
-					explosionDamage, 0f, explosionRadius, false);
+		{
+			if(justKnockback)
+			{
+				getWorld().getOtherEntities(null, new Box(getPos()).expand(explosionRadius), e -> true).forEach(e -> {
+					float dist = (float)getPos().toCenterPos().distanceTo(e.getPos());
+					e.addVelocity(e.getPos().subtract(getPos().toCenterPos()).normalize().multiply((1f - (dist / explosionRadius)) * explosionDamage));
+				});
+			}
+			else
+				ExplosionHandler.explosion(null, world, getPos().toCenterPos(), DamageSources.get(world, DamageSources.EXPLOSION),
+						explosionDamage, 0f, explosionRadius, false);
+		}
 		super.onStateChanged(newState);
 	}
 	
@@ -71,6 +84,7 @@ public class ExplosionListenerBlockEntity extends AbstractListenerBlockEntity
 			case "delay" -> activationDelay = Integer.parseInt(value);
 			case "damage" -> explosionDamage = Float.parseFloat(value);
 			case "radius" -> explosionRadius = Float.parseFloat(value);
+			case "justKnockback" -> justKnockback = Boolean.parseBoolean(value);
 		}
 		super.setAttribute(s, value);
 	}
@@ -83,6 +97,7 @@ public class ExplosionListenerBlockEntity extends AbstractListenerBlockEntity
 			case "delay" -> String.valueOf(activationDelay);
 			case "damage" -> String.valueOf(explosionDamage);
 			case "radius" -> String.valueOf(explosionRadius);
+			case "justKnockback" -> String.valueOf(justKnockback);
 			default -> null;
 		};
 	}
@@ -95,6 +110,8 @@ public class ExplosionListenerBlockEntity extends AbstractListenerBlockEntity
 			explosionDamage = nbt.getFloat("damage");
 		if(nbt.contains("radius", NbtElement.FLOAT_TYPE))
 			explosionRadius = nbt.getFloat("radius");
+		if(nbt.contains("justKnockback", NbtElement.BYTE_TYPE))
+			justKnockback = nbt.getBoolean("justKnockback");
 	}
 	
 	@Override
@@ -103,11 +120,13 @@ public class ExplosionListenerBlockEntity extends AbstractListenerBlockEntity
 		super.writeNbt(nbt);
 		nbt.putFloat("damage", explosionDamage);
 		nbt.putFloat("radius", explosionRadius);
+		nbt.putBoolean("justKnockback", justKnockback);
 	}
 	
 	static {
 		attributes.add("delay");
 		attributes.add("damage");
 		attributes.add("radius");
+		attributes.add("justKnockback");
 	}
 }