+++ src/main/java/absolutelyaya/ultracraft/client/rendering/entity/feature/gecko/V2EmissiveLayer.java
@@ -2,10 +2,7 @@ package absolutelyaya.ultracraft.client.rendering.entity.feature.gecko;
 
 import absolutelyaya.ultracraft.Ultracraft;
 import absolutelyaya.ultracraft.entity.machine.V2Entity;
-import net.minecraft.client.render.OverlayTexture;
-import net.minecraft.client.render.RenderLayer;
-import net.minecraft.client.render.VertexConsumer;
-import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.render.*;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.util.Identifier;
 import mod.azure.azurelib.cache.object.BakedGeoModel;
@@ -27,7 +24,7 @@ public class V2EmissiveLayer extends GeoRenderLayer<V2Entity>
 	@Override
 	public void render(MatrixStack poseStack, V2Entity animatable, BakedGeoModel bakedModel, RenderLayer renderType, VertexConsumerProvider bufferSource, VertexConsumer buffer, float partialTick, int packedLight, int packedOverlay)
 	{
-		RenderLayer armorRenderType;
+		RenderLayer renderLayer;
 		Identifier tex = switch((animatable.getMovementMode()) % 4)
 		{
 			default -> YELLOW;
@@ -35,9 +32,9 @@ public class V2EmissiveLayer extends GeoRenderLayer<V2Entity>
 			case 2 -> RED;
 			case 3 -> GREEN;
 		};
-		armorRenderType = RenderLayer.getEntityTranslucentEmissive(tex);
-		getRenderer().reRender(getDefaultBakedModel(animatable), poseStack, bufferSource, animatable, armorRenderType,
-				bufferSource.getBuffer(armorRenderType), partialTick, packedLight, OverlayTexture.DEFAULT_UV,
+		renderLayer = RenderLayer.getEntityCutout(tex);
+		getRenderer().reRender(getDefaultBakedModel(animatable), poseStack, bufferSource, animatable, renderLayer,
+				bufferSource.getBuffer(renderLayer), partialTick, LightmapTextureManager.MAX_BLOCK_LIGHT_COORDINATE, OverlayTexture.DEFAULT_UV,
 				1, 1, 1, 1);
 	}
 }